Add Ethereum side for Set Future Yield

+++ ethereum/contracts/ICash.sol
@@ -22,7 +22,7 @@ interface IERC20 {
  */
 interface ICash is IERC20 {
     function mint(address account, uint128 principal) external returns (uint);
-    function burn(address account, uint256 amount) external returns (uint128);
+    function burn(address account, uint amount) external returns (uint128);
     function setFutureYield(uint128 nextYield, uint128 nextIndex, uint nextYieldStartAt) external;
     function getCashIndex() external view returns (uint);
 }
